+++ experiment_utils/conf/data/batch_tfms/RandomErasing.yaml
@@ -1,4 +1,4 @@
-_target_: experiment_utils.utils_hydra.load_args_list
+_target_: experiment_utils.utils.hydra_utils.load_args_list
 
 defaults:
   - /data/transforms/RandomErasing@1
