+++ experiment_utils/conf/data/item_tfms/RandomResizedCrop_FlipItem.yaml
@@ -1,4 +1,4 @@
-_target_: experiment_utils.utils_hydra.load_args_list
+_target_: experiment_utils.utils.hydra_utils.load_args_list
 defaults:
     # - empty
   - /data/transforms/RandomResizedCrop@1
