+++ experiment_utils/conf/opt_fn/ranger.yaml
@@ -1,4 +1,4 @@
-_target_: experiment_utils.utils_hydra.load_partial_func
+_target_: experiment_utils.utils.hydra_utils.load_partial_func
 function: ranger.Ranger
 lr: 0.001
 # Ranger options
