+++ experiment_utils/conf/opt_fn/ranger_fastai.yaml
@@ -1,4 +1,4 @@
-_target_: experiment_utils.utils_hydra.load_partial_func
+_target_: experiment_utils.utils.hydra_utils.load_partial_func
 function: fastai.basics.ranger
 mom: 0.95
 sqr_mom: 0.99
